+++ Engine/ac/draw.h
@@ -157,7 +157,7 @@ void putpixel_scaled(Common::Bitmap *ds, int x, int y, int col);
 // Render game on screen
 void render_to_screen();
 void GfxDriverOnInitCallback(void *data);
-bool GfxDriverSpriteEvtCallback(int evt, int data);
+bool GfxDriverSpriteEvtCallback(int evt, intptr_t data);
 
 // Create the actsps[objid] image with the object drawn correctly.
 // Returns true if nothing at all has changed and actsps is still
